style the contact form

+++ peByShazia/shop/models.py
@@ -5,7 +5,7 @@ class products(models.Model):
     image=models.ImageField(upload_to='media')
     name=models.CharField(max_length=20)
     description=models.CharField(max_length=50)
-    price=models.PositiveBigIntegerField(null=False)
+    price=models.PositiveIntegerField(null=False)
 
     class Meta:
         verbose_name='product'
